LayoutTests/platform/mac-wk1/fast/scrolling/scroll-animator-basic-events-expected.txt

CONSOLE MESSAGE: MainLocalFrameView: mouseEnteredContentArea
CONSOLE MESSAGE: MainLocalFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: RenderLayer: mouseEnteredContentArea
CONSOLE MESSAGE: RenderLayer: mouseEnteredVerticalScrollbar
CONSOLE MESSAGE: MainLocalFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: RenderLayer: mouseMovedInContentArea
CONSOLE MESSAGE: RenderLayer: mouseIsDownInVerticalScrollbar
CONSOLE MESSAGE: RenderLayer: mouseIsUpInVerticalScrollbar
CONSOLE MESSAGE: MainLocalFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: RenderLayer: mouseMovedInContentArea
CONSOLE MESSAGE: RenderLayer: mouseExitedVerticalScrollbar
CONSOLE MESSAGE: MainLocalFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: RenderLayer: mouseExitedContentArea
This is a scrollable div.

PASS successfullyParsed is true

TEST COMPLETE
